+++ ERP.API/Models/CustomerUpdateViewModelJson.cs
@@ -17,7 +17,7 @@ namespace ERP.API.Models
         [StringLength(45)]
         public string cu_fullname { get; set; }
 
-        public byte? cu_type { get; set; }
+        public int? cu_type { get; set; }
 
         [StringLength(250)]
         public string cu_note { get; set; }
